+++ src/ViewInstance.js
@@ -221,7 +221,7 @@ class ViewInstance extends React.Component {
     this.redirectToQuickMarcPage('duplicate');
   };
 
-  selectInstanse = (selectedInstance) => {
+  selectInstance = (selectedInstance) => {
     const { history, location, match } = this.props;
     const instanceId = match.params.id;
 
@@ -233,43 +233,6 @@ class ViewInstance extends React.Component {
 
   toggleItemsMovement = () => {
     this.setState((prevState) => ({ isItemsMovement: !prevState.isItemsMovement }));
-  };
-
-  moveItems = (toHolding, items) => {
-    const { mutator } = this.props;
-    return mutator.movableItems.POST({
-      toHoldingsRecordId: toHolding,
-      itemIds: items,
-    })
-      .then(({ nonUpdatedIds }) => {
-        const hasErrors = Boolean(nonUpdatedIds?.length);
-
-        const message = hasErrors ? (
-          <FormattedMessage
-            id="ui-inventory.moveItems.instance.items.error"
-            values={{ items: nonUpdatedIds.join(', ') }}
-          />
-        ) : (
-          <FormattedMessage
-            id="ui-inventory.moveItems.instance.items.success"
-            values={{ count: items.length }}
-          />
-        );
-        const type = hasErrors ? 'error' : 'success';
-
-        this.calloutRef.current.sendCallout({ type, message });
-      })
-      .catch(() => {
-        this.calloutRef.current.sendCallout({
-          type: 'error',
-          message: (
-            <FormattedMessage
-              id="ui-inventory.moveItems.instance.items.error.server"
-              values={{ items: items.join(', ') }}
-            />
-          ),
-        });
-      });
   };
 
   goBack = (e) => {
@@ -315,7 +278,11 @@ class ViewInstance extends React.Component {
 
   handleImportRecordModalSubmit = (args) => {
     this.setState({ isImportRecordModalOpened: false });
-    this.props.mutator.query.update({ _path: `/inventory/import/${this.props.match.params.id}`, xid: args.externalIdentifier });
+    this.props.mutator.query.update({
+      _path: `/inventory/import/${this.props.match.params.id}`,
+      xidtype: args.externalIdentifierType,
+      xid: args.externalIdentifier,
+    });
   }
 
   handleImportRecordModalCancel = () => {
@@ -501,7 +468,6 @@ class ViewInstance extends React.Component {
   render() {
     const {
       match: { params: { id, holdingsrecordid, itemid } },
-      referenceTables,
       stripes,
       onClose,
       paneWidth,
@@ -537,17 +503,13 @@ class ViewInstance extends React.Component {
           onClose={onClose}
           actionMenu={this.createActionMenuGetter(instance)}
           instance={instance}
-          referenceData={referenceTables}
           tagsEnabled={tagsEnabled}
         >
           {
             (!holdingsrecordid && !itemid) ?
               (
-                <MoveItemsContext
-                  moveItems={this.moveItems}
-                >
+                <MoveItemsContext>
                   <HoldingsListContainer
-                    referenceData={referenceTables}
                     instance={instance}
                     draggable={this.state.isItemsMovement}
                     droppable
@@ -577,19 +539,22 @@ class ViewInstance extends React.Component {
         {
           this.state.findInstancePluginOpened && (
             <InstancePlugin
-              onSelect={this.selectInstanse}
+              onSelect={this.selectInstance}
               onClose={this.toggleFindInstancePlugin}
               withTrigger={false}
             />
           )
         }
 
-        <ImportRecordModal
-          isOpen={this.state.isImportRecordModalOpened}
-          currentExternalIdentifier={undefined}
-          handleSubmit={this.handleImportRecordModalSubmit}
-          handleCancel={this.handleImportRecordModalCancel}
-        />
+        <IfInterface name="copycat-imports">
+          <ImportRecordModal
+            isOpen={this.state.isImportRecordModalOpened}
+            currentExternalIdentifier={undefined}
+            handleSubmit={this.handleImportRecordModalSubmit}
+            handleCancel={this.handleImportRecordModalCancel}
+            id={id}
+          />
+        </IfInterface>
       </>
     );
   }
@@ -631,7 +596,6 @@ ViewInstance.propTypes = {
   onClose: PropTypes.func,
   onCopy: PropTypes.func,
   paneWidth: PropTypes.string.isRequired,
-  referenceTables: PropTypes.object.isRequired,
   resources: PropTypes.shape({
     allInstanceItems: PropTypes.object.isRequired,
     allInstanceHoldings: PropTypes.object.isRequired,