+++ src/edit/items/ItemForm.js
@@ -36,7 +36,6 @@ import RepeatableField from '../../components/RepeatableField';
 import ElectronicAccessFields from '../electronicAccessFields';
 import { memoize, mutators } from '../formUtils';
 import { validateOptionalField } from '../../utils';
-import { RemoteStorageApiProvider } from '../../RemoteStorage';
 import { LocationSelectionWithCheck } from '../common';
 
 import styles from './ItemForm.css';
@@ -416,11 +415,7 @@ class ItemForm extends React.Component {
                     addLabel={<FormattedMessage id="ui-inventory.addFormerId" />}
                     template={[{
                       component: TextField,
-                      label: (
-                        <FormattedMessage id="ui-inventory.formerId">
-                          {(message) => message}
-                        </FormattedMessage>
-                      )
+                      label: <FormattedMessage id="ui-inventory.formerId" />,
                     }]}
                   />
                 </Col>
@@ -816,33 +811,31 @@ class ItemForm extends React.Component {
               onToggle={this.handleAccordionToggle}
               label={<FormattedMessage id="ui-inventory.location" />}
             >
-              <RemoteStorageApiProvider>
-                <Row>
-                  <Col sm={4}>
-                    <Field
-                      label={<FormattedMessage id="ui-inventory.permanentLocation" />}
-                      name="permanentLocation.id"
-                      id="additem_permanentlocation"
-                      component={LocationSelectionWithCheck}
-                      fullWidth
-                      marginBottom0
-                    />
-                  </Col>
-                  <Col sm={4}>
-                    <Field
-                      label={<FormattedMessage id="ui-inventory.temporaryLocation" />}
-                      name="temporaryLocation.id"
-                      id="additem_temporarylocation"
-                      component={LocationSelectionWithCheck}
-                      fullWidth
-                      marginBottom0
-                    />
-                  </Col>
-                </Row>
-                <Row>
-                  <RemoteStorageWarning />
-                </Row>
-              </RemoteStorageApiProvider>
+              <Row>
+                <Col sm={4}>
+                  <Field
+                    label={<FormattedMessage id="ui-inventory.permanentLocation" />}
+                    name="permanentLocation.id"
+                    id="additem_permanentlocation"
+                    component={LocationSelectionWithCheck}
+                    fullWidth
+                    marginBottom0
+                  />
+                </Col>
+                <Col sm={4}>
+                  <Field
+                    label={<FormattedMessage id="ui-inventory.temporaryLocation" />}
+                    name="temporaryLocation.id"
+                    id="additem_temporarylocation"
+                    component={LocationSelectionWithCheck}
+                    fullWidth
+                    marginBottom0
+                  />
+                </Col>
+              </Row>
+              <Row>
+                <RemoteStorageWarning />
+              </Row>
             </Accordion>
             <Accordion
               open={accordions.acc09}